+++ client/app/components/layouts/index.js
@@ -1,24 +1,24 @@
-import Activity_detail from './activity_detail'
+import Activity_detail from './activity_detail';
 import Activity from './activity';
-import Chat from './chat'
+import Chat from './chat';
+import Landing from './landing';
 import Notification from './notification';
 import Post from './post';
 import PostActivity from './postactivity';
 import Profile from './profile';
 import Search from './search';
 import Settings from './settings';
-import Landing from './landing'
 
 
 export {
     Activity_detail,
     Activity,
     Chat,
+    Landing,
     Notification,
     Post,
     PostActivity,
     Profile,
     Search,
-    Settings,
-    Landing
+    Settings
 }
